+++ tests/unit/test_spatial.py
@@ -4,6 +4,7 @@ from unittest import TestCase
 from unittest.mock import ANY, call, patch
 
 import numpy as np
+from harmony_service_lib.exceptions import NoDataException
 from harmony_service_lib.message import Message
 from netCDF4 import Dataset
 from numpy.testing import assert_array_equal
@@ -44,18 +45,49 @@ class TestSpatial(TestCase):
         Albers Conic Equal Area projection, in the Alaska region.
 
         """
-        harmony_message = Message({'subset': {'bbox': [-160, 68, -145, 70]}})
-        above_varinfo = VarInfoFromDmr('tests/data/ABoVE_TVPRM_example.dmr')
+        with self.subTest('Spatial request range within granule region'):
+            harmony_message = Message({'subset': {'bbox': [-160, 68, -145, 70]}})
+            above_varinfo = VarInfoFromDmr('tests/data/ABoVE_TVPRM_example.dmr')
 
-        self.assertDictEqual(
-            get_spatial_index_ranges(
-                {'/NEE', '/x', '/y', '/time'},
-                above_varinfo,
-                'tests/data/ABoVE_TVPRM_prefetch.nc4',
-                harmony_message,
-            ),
-            {'/x': (37, 56), '/y': (7, 26)},
-        )
+            self.assertDictEqual(
+                get_spatial_index_ranges(
+                    {'/NEE', '/x', '/y', '/time'},
+                    above_varinfo,
+                    'tests/data/ABoVE_TVPRM_prefetch.nc4',
+                    harmony_message,
+                ),
+                {'/x': (37, 56), '/y': (7, 26)},
+            )
+        with self.subTest('Spatial request outside granule'):
+            harmony_message = Message({'subset': {'bbox': [120, -70, 140, -68]}})
+            with self.assertRaises(NoDataException) as context:
+                get_spatial_index_ranges(
+                    {'/NEE', '/x', '/y', '/time'},
+                    above_varinfo,
+                    'tests/data/ABoVE_TVPRM_prefetch.nc4',
+                    harmony_message,
+                )
+            self.assertEqual(
+                context.exception.message,
+                "Spatial subset request outside supported dimension range for {'/NEE'}",
+            )
+        with self.subTest('Spatial request outside granule - multiple variables'):
+            harmony_message = Message({'subset': {'bbox': [120, -70, 140, -68]}})
+            with self.assertRaises(NoDataException) as context:
+                get_spatial_index_ranges(
+                    {'/NEE', '/lat', '/lon', '/x', '/y', '/time'},
+                    above_varinfo,
+                    'tests/data/ABoVE_TVPRM_prefetch.nc4',
+                    harmony_message,
+                )
+
+            self.assertIn(
+                "Spatial subset request outside supported dimension range for ",
+                context.exception.message,
+            )
+            self.assertIn('/NEE', context.exception.message)
+            self.assertIn('/lat', context.exception.message)
+            self.assertIn('/lon', context.exception.message)
 
     def test_get_spatial_index_ranges_projected_from_coordinates(self):
         """Ensure that correct index ranges can be calculated for a SMAP L3
@@ -97,6 +129,7 @@ class TestSpatial(TestCase):
                     38,
                 ),
             }
+
             index_ranges = get_spatial_index_ranges(
                 required_variables,
                 smap_varinfo,
@@ -134,6 +167,39 @@ class TestSpatial(TestCase):
                 ),
                 expected_index_ranges,
             )
+
+        with self.subTest('projection gridded with coordinates NoDataException'):
+            harmony_message = Message(
+                {'subset': {'bbox': [-179.9, -89.8, -179.8, -89.5]}}
+            )
+            smap_varinfo = VarInfoFromDmr(
+                'tests/data/SC_SPL3FTP_004.dmr',
+                'SPL3FTP',
+                'hoss/hoss_config.json',
+            )
+            prefetch_path = 'tests/data/SC_SPL3FTP_004_Polar_prefetch.nc4'
+            required_variables = {
+                '/Freeze_Thaw_Retrieval_Data_Polar/surface_flag',
+                '/Freeze_Thaw_Retrieval_Data_Polar/latitude',
+                '/Freeze_Thaw_Retrieval_Data_Polar/longitude',
+            }
+            with self.assertRaises(NoDataException) as context:
+                index_ranges = get_spatial_index_ranges(
+                    required_variables,
+                    smap_varinfo,
+                    prefetch_path,
+                    harmony_message,
+                )
+            expected_substrings = [
+                "Spatial subset request outside supported dimension range for",
+                "/Freeze_Thaw_Retrieval_Data_Polar/surface_flag",
+                "/Freeze_Thaw_Retrieval_Data_Polar/latitude",
+                "/Freeze_Thaw_Retrieval_Data_Polar/longitude",
+            ]
+
+            actual_message = context.exception.message
+            for substring in expected_substrings:
+                self.assertIn(substring, actual_message)
 
     def test_get_spatial_index_ranges_geographic(self):
         """Ensure that correct index ranges can be calculated for:
@@ -260,6 +326,73 @@ class TestSpatial(TestCase):
                 ),
                 {'/latitude': (5, 44), '/longitude': (160, 199)},
             )
+        with self.subTest('Geographic dimensions - one dimension out of range'):
+            harmony_message = Message(
+                {'subset': {'bbox': [-179.9, -89.8, -179.8, -89.5]}}
+            )
+            smap_varinfo = VarInfoFromDmr(
+                'tests/data/ATL16_prefetch.dmr',
+                'ATL16',
+                'hoss/hoss_config.json',
+            )
+            prefetch_path = 'tests/data/ATL16_prefetch.nc4'
+            required_variables = {
+                '/npolar_grid_lat',
+                '/npolar_grid_lon',
+                '/npolar_asr',
+            }
+            with self.assertRaises(NoDataException) as context:
+                get_spatial_index_ranges(
+                    required_variables,
+                    smap_varinfo,
+                    prefetch_path,
+                    harmony_message,
+                )
+            self.assertEqual(
+                context.exception.message,
+                "Spatial subset request outside supported dimension range for {'/npolar_grid_lat'}",
+            )
+
+        with self.subTest(
+            'NoDataException test for Geographic dimensions with - 2 dimensions out of range'
+        ):
+
+            test_file_name1 = f'{self.test_dir}/test1.nc'
+            harmony_message_outofrange = Message(
+                {'subset': {'bbox': [179.8, -44.9, 179.9, -22.9]}}
+            )
+
+            with Dataset(test_file_name1, 'w', format='NETCDF4') as test_file1:
+                test_file1.createDimension('latitude', size=90)
+                test_file1.createDimension('longitude', size=180)
+
+                test_file1.createVariable('latitude', float, dimensions=('latitude',))
+                test_file1['latitude'][:] = np.linspace(0, 89.5, 90)
+                test_file1['latitude'].setncatts({'units': 'degrees_north'})
+
+                test_file.createVariable('longitude', float, dimensions=('longitude',))
+                test_file['longitude'][:] = np.linspace(0.5, 179.5, 180)
+                test_file['longitude'].setncatts({'units': 'degrees_east'})
+
+                with self.assertRaises(NoDataException) as context:
+                    get_spatial_index_ranges(
+                        {'/latitude', '/longitude'},
+                        self.varinfo,
+                        test_file_name1,
+                        harmony_message_outofrange,
+                    )
+                self.assertIn(
+                    "Spatial subset request outside supported dimension range for ",
+                    context.exception.message,
+                )
+                self.assertIn(
+                    '/latitude',
+                    context.exception.message,
+                )
+                self.assertIn(
+                    '/longitude',
+                    context.exception.message,
+                )
 
     @patch('hoss.spatial.get_dimension_index_range')
     @patch('hoss.spatial.get_projected_x_y_extents')